Columns quoting proposal.
This is only a proposal to support column quoting in the SELECT and ORDER
parts of the select statement.
It lacks of tests, correct exception generation and I don'y know what else! :)

I don't know either the full impact on activerecord, even if I am using
inside a rails project without problems (at the moment).