Stop creating SQL literals for aliases.
PostgreSQL requires that the aliases aren't quoted, but MySQL and SQLite
don't seem to care at all, and it seems like we would do better to avoid
creating SqlLiteral nodes where we don't need to.

I ran the mysql/postgresql/sqlite3 tests for ActiveRecord against this
branch and everything passes. Of course, we all know that doesn't
necessarily mean anything.

This is tangentially related to ernie/squeel#246, but still doesn't
address the "reserved word" complaint there. Might make sense to have an
Alias node for special visitation vs the SqlLiteral one, perhaps. Or
maybe we just don't care if someone wants to alias things using reserved
words in pg, because that's just cray-cray.