Allow Fixnum values to be quoted
Currently `User.where(id: '10')` would result in query where the
value is integer 10 instead of string 10. That's because id column is
of type 'Integer`.

However `User.where(name: 10)` would not result in value in string even
though name column is of type 'String'.

This PR allows "Fixnum" to be quoted so that in ActiveRecord proper
quoting can be done based on column type.